Sound Length caching and music (jukebox and ambience) improvments (#33)
* Refactors ambience to a subsystem (#56723)
(and #59071)
Ambience is now in a subsystem, and plays every now and then without you having to move to a new area for it to play

* our dll is NOT 3.8. grabs 3.6 from shiptest since i know its accurate.

* cache length of jukebox sounds and makes it so you only need the name

* updates the rust_g version for uptime

* acctually bumps us to 3.9 because im still lazy and am sure tg is correct.
based on https://github.com/tgstation/tgstation/pull/91121

* ambience improvments, mostly sound caching and less macros (#4599)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->
bits and pieces of
https://github.com/DaedalusDock/daedalusdock/pull/107
https://github.com/DaedalusDock/daedalusdock/pull/1123

Ambience is now stored in a big global that caches all of its sound
lengths on init instead of a bunch of macros.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Should stop ambeince from rarely cutting into eachother. Also areas wont
each create there own list of sounds so im pretty sure its a good
optimization.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

:cl:
fix: ambience should no longer have any chance of overlapping eachother
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

* play the beach sound lol

* move this

* improvments, mainly moving music to act like abience

* re-areas some of runtime town so i can test area changes better

* dont remove clients from the list just cause they are a newplayer

* longer delay while its still hardcodeed

* moves it off to its own prefrence and

* makes the beach ambience better looping

* comments out this office interior cause its annoying lol

* makes sure the generic one doesnt play in the vtm areas

---------

Co-authored-by: Qustinnus <Floydje123@hotmail.com>