Lupus Wolfkin Update (#62)
* update to dog.dm

Changes the lupus wolfkin to be its own simple mob, using the updated wolf sprites.

This means you will no longer revert a Lupus Wolfkin to 'corgi' when you attempt to feed or interact with them.

* Update SanFrancisco.dmm

* Changing them off var-edits

What it says on the tin, changed them off var-edits and made them individual in the list for ease of mapping.

* updated icon_dead

What it says on the tin, until we can get the damage overlay working on them, at least they will have a death icon to interact with.

* White wolfkin swapped

Still had the original base wolfkin in the map, that has been changed.

* Blood/Damage Overlay works

What it says on the tin, just making sure the Lupus Wolfkin will take damage properly and showcase they have taken that damage.

Right now its at set thresholds of brute, which is how its calculated on all mobs including players, in the future could try and set it to go based on percentile of HP.